Commit libxslt-1.1.29.tar.gz
This is the only source bundle we have to download over FTP...
Setting http_proxy doesn't work to download this resource.
I'm sure there's a way to fix this without committing this.

On the other hand, we could check in all source tar balls
in case any dependency is unavailable.